Add `Testcase.reducible` and simplify `Testcase.split_parts`.
Add `reducible` to the Testcase base class, which is used to avoid
removing parts of the testcase within DDBEGIN/END. This is used by
`jsstr` to make sure the splitter targets only the chars for reduction
in that round.

`Testcase.split_parts` now takes the entire testcase (not lines), and
adds parts directly into `self.parts` instead of yielding them to
`Testcase.load`.  This simplifies complex testcase splitters that need
state between lines.